Rename notes in note for note.py
