+++ app/api/v1/endpoints/note.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 from starlette import status
 
 from app.models.note import Note
-from app.schemas.notes import NoteOut, NoteCreate, NoteUpdate
+from app.schemas.note import NoteOut, NoteCreate, NoteUpdate
 from app.utils.dependency import get_db
 
 router = APIRouter()
